Ref: Substituting the badge domain file.
